pyproject.toml
[project]
name = "ai-chat-checkpoint"
version = "0.1.0"
description = "AI Chat Agent with Time Travel (Checkpoint) Feature using LangGraph and AzureChatOpenAI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Developer", email = "developer@example.com" }
]
keywords = ["langgraph", "langchain", "azure-openai", "chatbot", "checkpoint", "dash"]

dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "tavily-python>=0.3.0",
    "langfuse>=2.0.0",
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "python-dotenv>=1.0.0",
    "langchain-tavily>=0.2.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
chat-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
